Extract the repetitive labels
